WIP: remotenet-deploy-from-testnet.
