+++ package.json
@@ -41,7 +41,6 @@
     "setup-testnet": "cd testnet && python3 setup_testnet.py 1",
     "test-hbbft-contract": "NODE_ENV=localhost ts-node src/runHbbftTests.ts",
     "start-testnode": "cd testNode && ./start_node.sh",
-    "buildAndDeployNetwork": "ts-node src/runBuildAndDeployNetwork.ts 7",
     "typechain": "typechain --target web3-v1 --outDir src/abi/contracts 'src/abi/json/*.json'",
     "copy-abis": "mkdir -p src/abi/json/ && cp ../hbbft-posdao-contracts/build/contracts/*.json src/abi/json/",
     "logEvents": "ts-node src/runLogAvailabilityEvents.ts",
@@ -61,7 +60,8 @@
     "testnet-clear-validator-dbs": "ts-node src/regression/runClearValidatorsDB.ts",
     "testnet-fresh-grow-each-epoch": "npm-run-all testnet-fresh addOneValidatorEachEpoch",
     "testnet-fresh": "export STAKING_TRANSITION_WINDOW_LENGTH=90 && export STAKING_EPOCH_DURATION=120 && export MINIMUM_BLOCK_TIME=1 && cd testnet && ./setup_testnet.py 1 17 && cd -",
-    "realnet-fresh-1hour": "export STAKING_TRANSITION_WINDOW_LENGTH=600 && export STAKING_EPOCH_DURATION=3600 && export MINIMUM_BLOCK_TIME=5 && export MAXIMUM_BLOCK_TIME=60 && cd testnet && ./setup_testnet.py 1 7 && cd -"
+    "testnet-fresh-1hour": "export STAKING_TRANSITION_WINDOW_LENGTH=600 && export STAKING_EPOCH_DURATION=3900 && export MINIMUM_BLOCK_TIME=10 && export MAXIMUM_BLOCK_TIME=600 && cd testnet && ./setup_testnet.py 1 27 && cd -",
+    "remotenet-deploy-from-testnet": "ts-node src/remotenet/runDeployFromTestnet"
   },
   "repository": {
     "type": "git",
