+++ src/regression/nodeManager.ts
@@ -289,8 +289,6 @@ export class NodeManager {
     }
   }
 
-
-
 }
