+++ testnet/.gitignore
@@ -1,2 +1,3 @@
 
 nodes-current
+nodes-backup
